Rethinking and restructuring the ory/examples repo
Purpose: The community pages community.ory.sh indicate an increased interest in practical examples using the Ory Ecosystem in numerous languages including Go, Python, JS, Java, and Scala to name a few. The goal is to produce a set of oss examples that devs can use, copy and paste, to demonstrate the advantages of Ory in their own work.
PR: This PR makes some initial corrections to the example descriptions.